refactor(button): splits out button into versioned implementations
