+++ packages/palette/src/elements/Button/index.ts
@@ -1,2 +1,2 @@
 export * from "./Button"
-export * from "./tokens"
+export * from "./types"
