--- packages/palette/src/elements/Button/tokens/index.ts
@@ -1,2 +1,0 @@
-export * from "./v2"
-export * from "./v3"
